# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MCQ Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /subjects/{subjectId}/chapters:
    get:
      operationId: getChapters
      tags: [chapters]
      summary: List chapters for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a chapter (admin)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterInput"
      responses:
        "201":
          description: Chapter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{id}:
    get:
      operationId: getChapter
      tags: [chapters]
      summary: Get a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    patch:
      operationId: updateChapter
      tags: [chapters]
      summary: Update a chapter (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterUpdate"
      responses:
        "200":
          description: Chapter updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete a chapter (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # MCQs
...
  /bookmarks/{mcqId}:
    post:
      operationId: addBookmark
      tags: [bookmarks]
      summary: Bookmark an MCQ
      parameters:
        - name: mcqId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bookmarked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: removeBookmark
      tags: [bookmarks]
      summary: Remove bookmark
      parameters:
        - name: mcqId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Wrong answers
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban or unban a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanInput"
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth schemas
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        password:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, isPremium, isBanned, streakDays, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, admin]
        isPremium:
          type: boolean
        isBanned:
          type: boolean
        streakDays:
          type: integer
        totalPoints:
          type: integer
        createdAt:
          type: string
...
    Chapter:
      type: object
      required: [id, subjectId, name, description, mcqCount, order, isActive]
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        name:
          type: string
        description:
          type: string
        mcqCount:
          type: integer
        order:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
    ChapterInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        order:
          type: integer
    ChapterUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        order:
          type: integer
        isActive:
          type: boolean

    # MCQ schemas
...
    UserListResponse:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [student, admin]
        isPremium:
          type: boolean
        isBanned:
          type: boolean
    BanInput:
      type: object
      required: [banned]
      properties:
        banned:
          type: boolean
...