# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MCQ Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatar?: string | null;
  role: UserRole;
  isPremium: boolean;
  isBanned: boolean;
  streakDays: number;
  totalPoints?: number;
  createdAt: string;
}